Update to File Structure
