+++ nkedu-back/src/main/java/com/nkedu/back/api/SchoolService.java
@@ -2,7 +2,7 @@ package com.nkedu.back.api;
 
 import java.util.List;
 
-import com.nkedu.back.model.SchoolDTO;
+import com.nkedu.back.dto.SchoolDTO;
 
 public interface SchoolService {
