+++ nkedu-back/src/main/java/com/nkedu/back/dto/ParentDto.java
@@ -1,4 +1,4 @@
-package com.nkedu.back.model;
+package com.nkedu.back.dto;
 
 import java.sql.Date;
 
@@ -15,22 +15,19 @@ import lombok.Setter;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class StudentDTO {
+public class ParentDto {
 	
 	private Long id;
 	
 	private String username;
-	
-	@JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
-	private String password;
 
+	@JsonProperty(access = JsonProperty.Access.WRITE_ONLY) // 쓰기 전용
+	private String password;
+	
 	private String nickname;
 	
 	private Date birth;
 	
 	private String phoneNumber;
 	
-	private School school;
-	
-	private Long grade;
 }
