+++ nkedu-back/src/main/java/com/nkedu/back/dto/StudentDTO.java
@@ -1,9 +1,10 @@
-package com.nkedu.back.model;
+package com.nkedu.back.dto;
 
 import java.sql.Date;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import com.nkedu.back.entity.School;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -15,19 +16,22 @@ import lombok.Setter;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class ParentDto {
+public class StudentDTO {
 	
 	private Long id;
 	
 	private String username;
-
-	@JsonProperty(access = JsonProperty.Access.WRITE_ONLY) // 쓰기 전용
-	private String password;
 	
+	@JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+	private String password;
+
 	private String nickname;
 	
 	private Date birth;
 	
 	private String phoneNumber;
 	
+	private School school;
+	
+	private Long grade;
 }
