+++ nkedu-back/src/main/java/com/nkedu/back/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.nkedu.back.model;
+package com.nkedu.back.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
