+++ nkedu-back/src/main/java/com/nkedu/back/repository/ParentRepository.java
@@ -3,7 +3,7 @@ package com.nkedu.back.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.nkedu.back.model.Parent;
+import com.nkedu.back.entity.Parent;
 
 @Repository
 public interface ParentRepository extends JpaRepository<Parent, Long> {
