+++ nkedu-back/src/main/java/com/nkedu/back/repository/SchoolRepository.java
@@ -4,7 +4,7 @@ package com.nkedu.back.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.nkedu.back.model.School;
+import com.nkedu.back.entity.School;
 
 @Repository
 public interface SchoolRepository extends JpaRepository<School, Long> {
