+++ nkedu-back/src/main/java/com/nkedu/back/repository/StudentRepository.java
@@ -2,9 +2,8 @@ package com.nkedu.back.repository;
 
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-import com.nkedu.back.model.Student;
+import com.nkedu.back.entity.Student;
 
 public interface StudentRepository extends JpaRepository<Student, Long> {
