+++ nkedu-back/src/main/java/com/nkedu/back/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.nkedu.back.model.User;
+import com.nkedu.back.entity.User;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
